--- src/main/java/com/hhplus/concert_ticketing/infra/redis/RedisConfig.java
@@ -1,17 +1,0 @@
-package com.hhplus.concert_ticketing.infra.redis;
-
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RedisConfig {
-    @Bean
-    public RedissonClient redissonClient() {
-        Config config = new Config();
-        config.useSingleServer().setAddress("redis://localhost:6379");
-        return Redisson.create(config);
-    }
-}
