--- src/main/resources/application.properties
@@ -1,9 +1,0 @@
-spring.application.name=concert-ticketing
-spring.datasource.url=jdbc:h2:tcp://localhost/~/test
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-# JPA
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-#spring.jpa.hibernate.ddl-auto=none
-spring.jpa.hibernate.ddl-auto=create
